Implement user management and metrics endpoints
- Refactor user management endpoints in `users.py` to streamline user search and details retrieval.
- Introduce new metrics helper utilities in `_metrics_helpers.py` for shaping payloads and filtering activities.
- Add Discord data proxy endpoints in `guilds_discord.py` for roles, channels, and members management.
- Create organization settings and validation endpoints in `guilds_organization.py` for managing guild organization settings.
- Develop bulk ID resolution and CSV export endpoints in `users_bulk_export.py` for enhanced user data handling.
