+++ web/backend/routes/guilds.py
@@ -65,27 +65,15 @@ from core.schemas import (
     BotChannelSettings,
     BotChannelSettingsResponse,
     BotRoleSettings,
-    DiscordChannel,
-    DiscordRole,
     EventModuleSettings,
-    GuildChannelsResponse,
     GuildConfigData,
     GuildConfigResponse,
     GuildConfigUpdateRequest,
     GuildInfo,
     GuildInfoResponse,
-    GuildMember,
-    GuildMemberResponse,
-    GuildMembersResponse,
-    GuildRolesResponse,
-    LogoValidationRequest,
-    LogoValidationResponse,
     MetricsSettings,
     NewMemberRoleSettings,
     OrganizationSettings,
-    OrganizationSettingsResponse,
-    OrganizationValidationRequest,
-    OrganizationValidationResponse,
     ReadOnlyYamlConfig,
     RoleDelegationPolicy,
     UserProfile,
@@ -93,119 +81,14 @@ from core.schemas import (
 )
 from core.validation import (
     ensure_guild_match,
-    safe_int,
 )
-from fastapi import APIRouter, Depends, HTTPException, Query
+from fastapi import APIRouter, Depends, HTTPException
 
 if TYPE_CHECKING:
     from config.config_loader import ConfigLoader
 
 router = APIRouter(prefix="/api/guilds", tags=["guilds"])
 logger = logging.getLogger(__name__)
-
-
-def _coerce_role(role_data: dict) -> DiscordRole | None:
-    role_id = role_data.get("id")
-    if role_id is None:
-        return None
-
-    # Convert to string to preserve 64-bit Discord snowflake precision
-    role_id_str = str(role_id)
-    name = role_data.get("name") or "Unnamed Role"
-    color_value = role_data.get("color")
-    color = safe_int(color_value) if color_value is not None else None
-
-    return DiscordRole(
-        id=role_id_str,
-        name=str(name),
-        color=color,
-    )
-
-
-def _coerce_member(member_data: dict) -> GuildMember | None:
-    user_id = member_data.get("user_id")
-    user_id_int = safe_int(user_id)
-    if user_id_int is None:
-        return None
-
-    roles: list[DiscordRole] = []
-    for role in member_data.get("roles", []):
-        if not isinstance(role, dict):
-            continue
-        role_obj = _coerce_role(role)
-        if role_obj:
-            roles.append(role_obj)
-
-    return GuildMember(
-        user_id=user_id_int,
-        username=member_data.get("username"),
-        discriminator=member_data.get("discriminator"),
-        global_name=member_data.get("global_name"),
-        avatar_url=member_data.get("avatar_url"),
-        joined_at=member_data.get("joined_at"),
-        created_at=member_data.get("created_at"),
-        roles=roles,
-    )
-
-
-@router.get(
-    "/{guild_id}/roles/discord",
-    response_model=GuildRolesResponse,
-    dependencies=[Depends(require_fresh_guild_access)],
-)
-async def get_discord_roles(
-    guild_id: int,
-    current_user: UserProfile = Depends(require_staff()),
-    internal_api: InternalAPIClient = Depends(get_internal_api_client),
-):
-    """Return roles for a guild sourced from the bot's internal API."""
-    ensure_guild_match(guild_id, current_user)
-    try:
-        roles_payload = await internal_api.get_guild_roles(guild_id)
-    except Exception as exc:  # pragma: no cover - transport errors
-        raise translate_internal_api_error(exc, "Failed to fetch roles") from exc
-
-    roles: list[DiscordRole] = []
-    for role in roles_payload:
-        role_obj = _coerce_role(role)
-        if role_obj:
-            roles.append(role_obj)
-
-    return GuildRolesResponse(roles=roles)
-
-
-@router.get(
-    "/{guild_id}/channels/discord",
-    response_model=GuildChannelsResponse,
-    dependencies=[Depends(require_fresh_guild_access)],
-)
-async def get_discord_channels(
-    guild_id: int,
-    current_user: UserProfile = Depends(require_staff()),
-    internal_api: InternalAPIClient = Depends(get_internal_api_client),
-):
-    """Return text channels for a guild sourced from the bot's internal API."""
-    ensure_guild_match(guild_id, current_user)
-    try:
-        channels_payload = await internal_api.get_guild_channels(guild_id)
-    except Exception as exc:  # pragma: no cover - transport errors
-        raise translate_internal_api_error(exc, "Failed to fetch channels") from exc
-
-    channels: list[DiscordChannel] = []
-    for channel in channels_payload:
-        channel_id = channel.get("id")  # Already a string from bot API
-        if channel_id:
-            channels.append(
-                DiscordChannel(
-                    id=str(channel_id),  # Ensure it's a string
-                    name=channel.get("name", "Unknown"),
-                    category=channel.get("category"),
-                    position=channel.get("position", 0),
-                    type=safe_int(channel.get("type")),
-                )
-            )
-
-    return GuildChannelsResponse(channels=channels)
 
 
 @router.get(
@@ -410,252 +293,6 @@ async def update_voice_selectable_roles_settings(
     updated = await get_voice_selectable_roles(db, guild_id)
     await _notify_refresh(internal_api, guild_id, source="voice_selectable_roles")
     return VoiceSelectableRoles(selectable_roles=updated)
-
-
-@router.get(
-    "/{guild_id}/members",
-    response_model=GuildMembersResponse,
-    dependencies=[Depends(require_fresh_guild_access)],
-)
-async def list_guild_members(
-    guild_id: int,
-    page: int = Query(1, ge=1, le=1000),
-    page_size: int = Query(100, ge=1, le=1000),
-    current_user: UserProfile = Depends(require_staff()),
-    internal_api: InternalAPIClient = Depends(get_internal_api_client),
-):
-    """Proxy paginated Discord member data from the bot's internal API."""
-    ensure_guild_match(guild_id, current_user)
-
-    try:
-        payload = await internal_api.get_guild_members(
-            guild_id,
-            page=page,
-            page_size=page_size,
-        )
-    except Exception as exc:
-        raise translate_internal_api_error(
-            exc,
-            "Failed to fetch guild members",
-        ) from exc
-
-    members: list[GuildMember] = []
-    for raw_member in payload.get("members", []):
-        member_obj = _coerce_member(raw_member)
-        if member_obj:
-            members.append(member_obj)
-
-    return GuildMembersResponse(
-        members=members,
-        page=payload.get("page", page),
-        page_size=payload.get("page_size", page_size),
-        total=payload.get("total", len(members)),
-    )
-
-
-@router.get(
-    "/{guild_id}/members/{user_id}",
-    response_model=GuildMemberResponse,
-    dependencies=[Depends(require_fresh_guild_access)],
-)
-async def get_guild_member_detail(
-    guild_id: int,
-    user_id: int,
-    current_user: UserProfile = Depends(require_staff()),
-    internal_api: InternalAPIClient = Depends(get_internal_api_client),
-):
-    """Proxy a single Discord member lookup via the internal API."""
-    ensure_guild_match(guild_id, current_user)
-
-    try:
-        payload = await internal_api.get_guild_member(guild_id, user_id)
-    except Exception as exc:
-        raise translate_internal_api_error(
-            exc,
-            "Failed to fetch guild member",
-        ) from exc
-
-    member = _coerce_member(payload)
-    if not member:
-        raise HTTPException(
-            status_code=502,
-            detail="Invalid member payload from internal API",
-        )
-
-    return GuildMemberResponse(member=member)
-
-
-# Organization Settings Endpoints
-
-
-# Load RSI config via centralized ConfigLoader (no direct file IO)
-def _load_rsi_config() -> dict:
-    """Return RSI configuration from the shared ConfigLoader cache."""
-    from core.dependencies import get_config_loader
-
-    loader = get_config_loader()
-    config = loader.load_config()
-    return config.get("rsi", {})
-
-
-# Initialize RSI client (lazy)
-_rsi_client = None
-
-
-def _get_rsi_client():
-    """Get or create RSI client singleton."""
-    global _rsi_client
-    if _rsi_client is None:
-        from core.rsi_utils import RSIClient
-
-        rsi_config = _load_rsi_config()
-        requests_per_minute = rsi_config.get("requests_per_minute", 30)
-        user_agent = rsi_config.get("user_agent", "TEST-Squadron-Verification-Bot/1.0")
-        _rsi_client = RSIClient(
-            requests_per_minute=requests_per_minute, user_agent=user_agent
-        )
-    return _rsi_client
-
-
-@router.get(
-    "/{guild_id}/settings/organization",
-    response_model=OrganizationSettings,
-    dependencies=[Depends(require_fresh_guild_access)],
-)
-async def get_organization_settings_endpoint(
-    guild_id: int,
-    db=Depends(get_db),
-    current_user: UserProfile = Depends(require_staff()),
-):
-    """Fetch organization settings for a guild."""
-    ensure_guild_match(guild_id, current_user)
-    settings = await get_organization_settings(db, guild_id)
-    return OrganizationSettings(**settings)
-
-
-@router.put(
-    "/{guild_id}/settings/organization",
-    response_model=OrganizationSettingsResponse,
-    dependencies=[Depends(require_fresh_guild_access)],
-)
-async def update_organization_settings_endpoint(
-    guild_id: int,
-    payload: OrganizationSettings,
-    db=Depends(get_db),
-    current_user: UserProfile = Depends(require_bot_admin()),
-    internal_api: InternalAPIClient = Depends(get_internal_api_client),
-):
-    """Update organization settings for a guild."""
-    ensure_guild_match(guild_id, current_user)
-
-    # Validate logo URL if provided
-    validated_logo_url = None
-    if payload.organization_logo_url:
-        try:
-            validated_logo_url = await validate_logo_url(payload.organization_logo_url)
-        except LogoValidationError as exc:
-            raise HTTPException(status_code=400, detail=str(exc)) from exc
-
-    # Get current settings to detect logo change
-    current_org = await get_organization_settings(db, guild_id)
-    logo_changed = current_org.get("organization_logo_url") != validated_logo_url
-
-    await set_organization_settings(
-        db,
-        guild_id,
-        payload.organization_sid,
-        payload.organization_name,
-        validated_logo_url,
-    )
-    updated = await get_organization_settings(db, guild_id)
-    await _notify_refresh(internal_api, guild_id, source="organization")
-
-    # Track verification message update status
-    verification_message_updated: bool | None = None
-
-    # Trigger verification message repost if logo changed
-    if logo_changed:
-        try:
-            await internal_api.resend_verification_message(guild_id)
-            verification_message_updated = True
-        except Exception as exc:
-            logger.warning(
-                "Failed to resend verification message for guild %s after logo update: %s",
-                guild_id,
-                exc,
-            )
-            verification_message_updated = False
-
-    return OrganizationSettingsResponse(
-        **updated,
-        verification_message_updated=verification_message_updated,
-    )
-
-
-@router.post(
-    "/{guild_id}/organization/validate-sid",
-    response_model=OrganizationValidationResponse,
-    dependencies=[Depends(require_fresh_guild_access)],
-)
-async def validate_organization_sid_endpoint(
-    guild_id: int,
-    payload: OrganizationValidationRequest,
-    current_user: UserProfile = Depends(require_bot_admin()),
-):
-    """Validate an organization SID by fetching from RSI."""
-    ensure_guild_match(guild_id, current_user)
-
-    from core.rsi_utils import validate_organization_sid
-
-    rsi_client = _get_rsi_client()
-
-    is_valid, org_name, error_msg = await validate_organization_sid(
-        payload.sid, rsi_client
-    )
-
-    return OrganizationValidationResponse(
-        success=True,
-        is_valid=is_valid,
-        sid=payload.sid.strip().upper(),
-        organization_name=org_name,
-        error=error_msg,
-    )
-
-
-@router.post(
-    "/{guild_id}/organization/validate-logo",
-    response_model=LogoValidationResponse,
-    dependencies=[Depends(require_fresh_guild_access)],
-)
-async def validate_logo_url_endpoint(
-    guild_id: int,
-    payload: LogoValidationRequest,
-    current_user: UserProfile = Depends(require_bot_admin()),
-):
-    """Validate a logo URL is reachable and returns an acceptable image.
-
-    Checks:
-    - URL is accessible (HEAD/GET request)
-    - Content-Type is an image type (png, jpg, gif, webp)
-    - File size is under 8MB
-    """
-    ensure_guild_match(guild_id, current_user)
-
-    try:
-        validated_url = await validate_logo_url(payload.url)
-        return LogoValidationResponse(
-            success=True,
-            is_valid=True,
-            url=validated_url,
-            error=None,
-        )
-    except LogoValidationError as exc:
-        return LogoValidationResponse(
-            success=True,
-            is_valid=False,
-            url=None,
-            error=str(exc),
-        )
 
 
 # Guild info (for header)
