+++ web/backend/routes/metrics.py
@@ -1,21 +1,23 @@
-"""
-Metrics endpoints for the dashboard.
-
-Provides aggregated user activity metrics: voice time, game tracking,
-message counts, leaderboards, and time-series data for charting.
-"""
+"""Metrics endpoints for dashboard activity and leaderboard data."""
 
 import asyncio
 import contextlib
 import logging
 import time
 
+from ._metrics_helpers import (
+    build_message_leaderboard_entries,
+    build_voice_leaderboard_entries,
+    normalize_leaderboard_entries,
+    normalize_timeseries_data,
+    resolve_guild_id,
+    resolve_activity_filter,
+)
 from core.dependencies import (
     InternalAPIClient,
     get_internal_api_client,
     require_discord_manager,
 )
-from core.pagination import is_all_guilds_mode
 from core.schemas import (
     ActivityGroupCounts,
     ActivityGroupCountsResponse,
@@ -24,7 +26,6 @@ from core.schemas import (
     GameMetrics,
     GameMetricsResponse,
     LeaderboardResponse,
-    MessageLeaderboardEntry,
     MetricsOverview,
     MetricsOverviewResponse,
     TimeSeriesResponse,
@@ -32,7 +33,6 @@ from core.schemas import (
     UserMetrics,
     UserMetricsResponse,
     UserProfile,
-    VoiceLeaderboardEntry,
 )
 from fastapi import APIRouter, Depends, HTTPException, Query
 
@@ -42,181 +42,8 @@ router = APIRouter()
 logger = logging.getLogger(__name__)
 MAX_GAME_DETAIL_FILTER_USERS = 1000
 
-
-def _resolve_guild_id(current_user: UserProfile) -> int:
-    """
-    Extract and validate the active guild ID from the current user session.
-
-    Raises HTTPException if no guild is selected or in cross-guild mode
-    (metrics are per-guild only).
-    """
-    if not current_user.active_guild_id or is_all_guilds_mode(
-        current_user.active_guild_id
-    ):
-        raise HTTPException(
-            status_code=400,
-            detail="Metrics require a specific guild selection. Cross-guild mode is not supported.",
-        )
-    return int(current_user.active_guild_id)
-
-
-async def _resolve_activity_filter(
-    internal_api: InternalAPIClient,
-    guild_id: int,
-    dimension: str | None,
-    tier: str | None,
-    days: int = 30,
-) -> list[int] | None:
-    """Resolve dimension+tier into a list of user IDs, or None if no filter.
-
-    Supports comma-separated dimensions and tiers; users are matched if they
-    belong to the selected tier in ANY selected dimension.
-
-    Uses the bulk internal endpoint to resolve all dimension×tier combos in a
-    single HTTP call instead of issuing one request per pair.
-    """
-    if not dimension or not tier:
-        return None
-    raw_dims = [part.strip() for part in dimension.split(",") if part.strip()]
-    if not raw_dims:
-        return None
-    resolved_dims: list[str] = []
-    for raw in raw_dims:
-        resolved = "combined" if raw == "all" else raw
-        if resolved not in resolved_dims:
-            resolved_dims.append(resolved)
-
-    raw_tiers = [part.strip() for part in tier.split(",") if part.strip()]
-    if not raw_tiers:
-        return None
-    resolved_tiers: list[str] = []
-    for raw in raw_tiers:
-        if raw not in resolved_tiers:
-            resolved_tiers.append(raw)
-
-    try:
-        bulk = await internal_api.get_activity_group_members_bulk(
-            guild_id,
-            resolved_dims,
-            resolved_tiers,
-            days=days,
-        )
-        merged_user_ids: set[int] = set()
-        for _dim_key, tier_map in bulk.items():
-            if not isinstance(tier_map, dict):
-                continue
-            for _tier_key, uid_list in tier_map.items():
-                if not isinstance(uid_list, list):
-                    continue
-                for uid in uid_list:
-                    try:
-                        merged_user_ids.add(int(uid))
-                    except (TypeError, ValueError):
-                        continue
-        return sorted(merged_user_ids)
-    except Exception as exc:
-        logger.warning(
-            "Failed to resolve activity filter",
-            exc_info=exc,
-        )
-        return []
-
-
-def _coerce_metric_value(raw_value: object) -> int | None:
-    """
-    Coerce leaderboard metric values to integers when possible.
-
-    AI Notes:
-        Leaderboard metrics are expected to be numeric counts/durations, but
-        upstream payloads may occasionally contain loosely typed values in the
-        generic ``value`` field. The boolean case is defensive normalization so
-        unexpected ``true``/``false`` values become ``1``/``0`` instead of
-        being rejected during response shaping.
-    """
-    if isinstance(raw_value, bool):
-        # Defensive handling for inconsistent upstream payloads; booleans are
-        # not an expected leaderboard metric type, but some serializers may
-        # emit them in the generic "value" field.
-        return int(raw_value)
-    if isinstance(raw_value, int):
-        return raw_value
-    if isinstance(raw_value, float):
-        return int(raw_value)
-    if isinstance(raw_value, str):
-        with contextlib.suppress(ValueError):
-            return int(float(raw_value))
-    return None
-
-
-def _normalize_leaderboard_entries(
-    entries: list[object],
-    *,
-    metric_field: str,
-) -> list[dict[str, object]]:
-    """Normalize leaderboard payloads for stable frontend consumption."""
-    normalized_entries: list[dict[str, object]] = []
-    for entry in entries:
-        if not isinstance(entry, dict):
-            continue
-        normalized = dict(entry)
-        if "user_id" in normalized:
-            normalized["user_id"] = str(normalized["user_id"])
-        if metric_field not in normalized and "value" in normalized:
-            coerced_value = _coerce_metric_value(normalized.get("value"))
-            if coerced_value is not None:
-                normalized[metric_field] = coerced_value
-        normalized.pop("value", None)
-        normalized_entries.append(normalized)
-    return normalized_entries
-
-
-def _normalize_timeseries_data(data: list[object]) -> list[dict[str, object]]:
-    """Filter timeseries payloads down to dict items for response stability."""
-    return [dict(item) for item in data if isinstance(item, dict)]
-
-
-def _build_voice_leaderboard_entries(
-    entries: list[object],
-) -> list[VoiceLeaderboardEntry]:
-    """Build typed voice leaderboard entries for the bundled response."""
-    typed_entries: list[VoiceLeaderboardEntry] = []
-    for entry in _normalize_leaderboard_entries(entries, metric_field="total_seconds"):
-        total_seconds = _coerce_metric_value(entry.get("total_seconds"))
-        if total_seconds is None:
-            continue
-        username = entry.get("username")
-        avatar_url = entry.get("avatar_url")
-        typed_entries.append(
-            VoiceLeaderboardEntry(
-                user_id=str(entry.get("user_id", "")),
-                total_seconds=total_seconds,
-                username=username if isinstance(username, str) else None,
-                avatar_url=avatar_url if isinstance(avatar_url, str) else None,
-            )
-        )
-    return typed_entries
-
-
-def _build_message_leaderboard_entries(
-    entries: list[object],
-) -> list[MessageLeaderboardEntry]:
-    """Build typed message leaderboard entries for the bundled response."""
-    typed_entries: list[MessageLeaderboardEntry] = []
-    for entry in _normalize_leaderboard_entries(entries, metric_field="total_messages"):
-        total_messages = _coerce_metric_value(entry.get("total_messages"))
-        if total_messages is None:
-            continue
-        username = entry.get("username")
-        avatar_url = entry.get("avatar_url")
-        typed_entries.append(
-            MessageLeaderboardEntry(
-                user_id=str(entry.get("user_id", "")),
-                total_messages=total_messages,
-                username=username if isinstance(username, str) else None,
-                avatar_url=avatar_url if isinstance(avatar_url, str) else None,
-            )
-        )
-    return typed_entries
+# Backward-compatible alias for tests and existing monkeypatch hooks.
+_resolve_activity_filter = resolve_activity_filter
 
 
 @router.get("/overview", response_model=MetricsOverviewResponse)
@@ -240,7 +67,7 @@ async def get_metrics_overview(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
     started_at = time.perf_counter()
 
     try:
@@ -274,7 +101,7 @@ async def get_dashboard_metrics(
     internal_api: InternalAPIClient = Depends(get_internal_api_client),
 ) -> DashboardMetricsResponse:
     """Get a bundled metrics payload optimized for the dashboard page."""
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
     started_at = time.perf_counter()
 
     try:
@@ -312,17 +139,17 @@ async def get_dashboard_metrics(
         return DashboardMetricsResponse(
             data=DashboardMetricsBundle(
                 overview=MetricsOverview(**overview_result),
-                voice_leaderboard=_build_voice_leaderboard_entries(
+                voice_leaderboard=build_voice_leaderboard_entries(
                     voice_result.get("entries", [])
                 ),
-                message_leaderboard=_build_message_leaderboard_entries(
+                message_leaderboard=build_message_leaderboard_entries(
                     message_result.get("entries", [])
                 ),
                 top_games=top_games_result.get("games", []),
-                message_timeseries=_normalize_timeseries_data(
+                message_timeseries=normalize_timeseries_data(
                     message_timeseries_result.get("data", [])
                 ),
-                voice_timeseries=_normalize_timeseries_data(
+                voice_timeseries=normalize_timeseries_data(
                     voice_timeseries_result.get("data", [])
                 ),
                 activity_counts=ActivityGroupCounts(**activity_groups_result),
@@ -358,7 +185,7 @@ async def get_voice_leaderboard(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
 
     try:
         user_ids = await _resolve_activity_filter(
@@ -367,7 +194,7 @@ async def get_voice_leaderboard(
         result = await internal_api.get_metrics_voice_leaderboard(
             guild_id, days=days, limit=limit, user_ids=user_ids
         )
-        normalized_entries = _normalize_leaderboard_entries(
+        normalized_entries = normalize_leaderboard_entries(
             result.get("entries", []), metric_field="total_seconds"
         )
         return LeaderboardResponse(entries=normalized_entries)
@@ -396,7 +223,7 @@ async def get_message_leaderboard(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
 
     try:
         user_ids = await _resolve_activity_filter(
@@ -405,7 +232,7 @@ async def get_message_leaderboard(
         result = await internal_api.get_metrics_message_leaderboard(
             guild_id, days=days, limit=limit, user_ids=user_ids
         )
-        normalized_entries = _normalize_leaderboard_entries(
+        normalized_entries = normalize_leaderboard_entries(
             result.get("entries", []), metric_field="total_messages"
         )
         return LeaderboardResponse(entries=normalized_entries)
@@ -434,7 +261,7 @@ async def get_top_games(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
 
     try:
         user_ids = await _resolve_activity_filter(
@@ -470,7 +297,7 @@ async def get_game_metrics(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
 
     try:
         user_ids = await _resolve_activity_filter(
@@ -533,7 +360,7 @@ async def get_timeseries(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
     started_at = time.perf_counter()
 
     try:
@@ -578,7 +405,7 @@ async def get_activity_groups(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
     started_at = time.perf_counter()
 
     try:
@@ -612,7 +439,7 @@ async def get_user_metrics(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
 
     try:
         result = await internal_api.get_metrics_user(guild_id, user_id, days=days)
@@ -642,7 +469,7 @@ async def delete_user_metrics(
 
     Requires: Discord Manager role or higher
     """
-    guild_id = _resolve_guild_id(current_user)
+    guild_id = resolve_guild_id(current_user)
     admin_user_id = int(current_user.user_id)
 
     try:
